requirements.txt
chromadb
requests
rich
pyyaml
